Bot
From the beginning, when the bot receives a picture, it determines the file type, if it is not .jpg, .jpeg, or .png, then the file is not accepted and the bot asks to give the file in the correct format. Next, the bot will save the image, process it, and determine its class.